uninstall: fix check if containers using kata exist
When we uninstall we check for running pods that still use kata-runtime.
If yes we ask the user to delete these pods before the uninstall
procedure continues.

There was a bug in the if condition that checked the RuntimeClass. In
the pod spec the RuntimeClassName is a reference (pointer) to a string
while in Kataconfig RuntimeClass is a string. The comparison was done
between two pointers not the string values of the runtime class name.

Fix this by 1. checkinf if the pointer in pod.spec is set and 2.
changing the if condition to de-reference the string in
pod.spec.RuntimeClassName and compare it to the one in Kataconfig.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>