Check compatibility of payload image and OCP version
This adds a check to make sure that the RPMs in the used
kata-operator-payload image are actually installable on the OCP version
used in the cluster.

To find out which version of OpenShift is installed we need to give
the service account the access rights to GET the clusterversion object.

Get the clusterversion and compare it to the tag of the kata payload
image specified in the kataconfig custom resource
(spec.config.sourceImage). If both versions match the installation
continues, otherwise reconcile will fail until a different image is
used.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>