add support for master/worker combined nodes
In some compact clusters variants nodes have both the master and worker
role.  However, the machine config pools that are used in
these configurations have all nodes in the master pool and zero nodes in
the worker pool.

Zero nodes in the worker pool breaks our current practice of installing
kata on all nodes in the worker pool. The solution is to install on all
nodes in the master pool instead. For that we fetch the worker pool
information and check if it has nodes in it. If it has a machine count
of zero we use the master pool instead.

This has an important implication:
in this master-only configuration we can't support installing on a
subset of nodes by setting a label on the selected nodes the user wants
to install kata on. Why? This feature depends on creating a custom pool
that inherits from the worker pool. MCP only supports inheriting from
the worker pool but not from the master pool. See
https://github.com/openshift/machine-config-operator/blob/master/docs/custom-pools.md

This is a trade-off for being able to support these compact cluster
types.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>