openshift: handle errors when updating kataconfig
It's perfectly normal when there is no 'kata-oc' machine config pool, so
don't treat is as an error and print a simple informational message
instead.

When adding the finalizer it's possible that we're trying to update
kataconfig but some other component has changed it after we fetched it
last. In this case return the error and try again.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>