Changes to make the kata install daemon work with operator sdk 1.0
When we started work on the kata-operator we used the Operator SDK v0.16
which is now very outdated and discontinued. For this reason we decided
to make the switch to SDK 1.0 and generate a new operator.

This had effects on the daemon part as well. It is now using a dynamic
kataconfig client. The part that installs the RPMs remains the same,
we still use the kata-operator-payload image to download the RPMs and
install them on the nodes host file system.

All of this is imported from github.com/harche/kata-operator-daemon
branch operator_sdk_upgrade_pr

Signed-off-by: Harshal Patil <harpatil@redhat.com>
Signed-off-by: Jens Freimann <jfreimann@redhat.com>