fix runtime class in example-fedora.yaml
We've switched to 'kata' as runtimeclass because it is what upstream
kata-deploy uses as well because users got confused about the -oc.

This is why we switched to 'kata' as the name for the 4.6 release already.

On the master branch we are still missing this change.
Fix the example pod file and the README accordingly.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>