use digests for kube-rbac-proxy and kata-operator-daemon images
In disconnected environments container images are mirrored to a local
system. A imageContentSourcePolicy resource needs to be created that
points to the mirrored image. Since OCP 4.3 the default behaviour of
ICSP is mirror-by-digest-only.

For ease of use in disconnected environment this patch switches the
references to container images in the operator code to use digests
instead of tags.

This patch is not changing the daemon to download the payload image via
digests on purpose. For disconnected environments the configmap for
payloads can be used.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>