add kata-operator-payload dockerfiles
This is a 1:1 copy from the release-4.6 branch
(commit 290bb0b3a139f51922b1ee0ebd0358edd6852784)

It builds a container image with the kata-* and QEMU (plus
dependencies) RPMs in it. The kata-operator-daemon when running
on the worker nodes downloads this image and then installs
the RPMs it contains.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>