controller: add environment variables for daemon secret
When using a private repository for the payload image the
daemon can now use the authentification credentials from a
Kubernetes secret that is provided by the user.

This patch adds two environment variables for username and password
to the daemon pod specification. The values are taken from a
secret that the user needs to create before the pods are created.
The environment variables are declared as optional so that no error
occurs without the secret.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>