daemon: use env variables to download payload from private registry
When a payload image is stored in a private repository the daemon
needs to authenticate with the registry to be able to download it.

There are two environment variables defined in the daemons pod
specification. These variables are populated by a Kubernetes secret that
the user can create. It needs to be created before the daemons start to
run.

The variables and the secret are optional and will only be used when
a configmap for a custom payload image is used. See the documentation
for instructions for how to use the configmap and the secret for custom
payload containers.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>