controller: The CRI-O drop-in file should be provided by the RPM
Instead of having the logic of the CRI-O drop-in file in the operator,
let's just rely on it being distributed by the `kata-containers` RPM.

This is done because that file only makes sense in a system where
`kata-containers` RPM is installed, because that file is only used in a
system where `kata-containers` RPM is installed.  So, leave it for the
RPM to take care of that and let's make our lives simpler on the
operator side.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>