controller: The service should ne enabled by the OS
`kata-osbuilder-generate.service` should be part of the RHCOS preset,
making it "enabled" by default whenever it's present / installed.

With that in mind, we can remove this logic from the operator and rely
on the OS to do the right thing for us.

Please, for more information about the work done on RHCOS side, refer
to: https://bugzilla.redhat.com/show_bug.cgi?id=1941342

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>