enable sandboxed-containers extension
Until now we used a separate image containing RPMs that were installed
by a daemon running on each node. This was error prone and complicated.
With this patch we can simplify the operator overall because we don't
need a daemon and payload image any more.

This patch changes the way Kata, QEMU and dependencies are installed in
the cluster. Instead of installing RPMs directly the operator now
enables a RHCOS extension using the Machine Config Operator. This is
achieved by creating a machine config object that enables the 'sandboxed
containers' extension.

This simplifies the installation of the runtime quite a lot but it also
brings some necessary changes with it. Before this patch the daemon was
updating the Kataconfig CR with status information about the
installation. This has to change now because there is no daemon any
more. Instead these status information is filled in by the controller.
The required information is gathered from the Machine Config Pool and
Node objects and is displayed in a very similar manner as before in the
description of the Kataconfig CR.

This brings no changes in how the operator itself is deployed and the
installation of the Kata runtime is triggered. By creating an instance
of the kataconfig CR the installation is triggered. The most important
part is the creation of the machine config object that enables the
sandboxed-containeres RHCOS extension. The controller then monitors the
progress by watching the data in the description of the machine config
pool. Once all nodes are updated with the new machine config state the
final step is to create the runtime class named 'kata'.

Uninstallation of the Kata runtime works in the same way. The machine
config entry is removed and this triggers MCO to disable the RHCOS
extension and eventually remove the kata runtime class.

Support for 3-node clusters stays the same as well as the possibility to
install the runtime on a chosen subset of the worker nodes by adding a
label to each node.

This patch also removes all files to build the daemon and payload image.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>