use new namespace name and refer to new quay.io repository
We should adhere to the naming convention in OpenShift and use the
default namespace openshift-sandboxed-containers-operator. The same goes
for the quay.io user/repository we use. Use the new
quay.io/openshift_sandboxed_containers instead of
quay.io/isolatedcontainres.

This patch does two things:
1. it cleans up README and example files to use the namespace
openshift-sandboxed-containers instead of
sandboxed-containers-operator-system.
2. use version 1.0 instead of 4.8

I've disabled namespacePrefix in config/default/kustomize.yaml to avoid
it to add "-system" to the namespace. This was the only way to make it
work since kustomize doesn't allow to leave the name field in the
namespace metadata blank.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>