openshift controller: use upgradeable condition
There are situations when we don't want OLM to initiate an
update/upgrade of the Operator:
 - during an installation of Kata Containers
 - during an uninstallation of Kata Containers

In these situations the state of Kata on the worker nodes is
in-flight and an update/upgrade could result in inconsistent state

OLM has built-in support for this and checks if the Operator has a
condition called 'Upgradeable'. OLM will not attempt an update when the
Operator's Upgradeable condition is set to false.
https://olm.operatorframework.io/docs/advanced-tasks/communicating-operator-conditions-to-olm/#upgradeable

We set the condition to *false* when
 - the kataconfig CR was created and we initiated the installation
 - the kataconfig Cr was deleted and we initiated the uninstllation

We set the condition to *true* when
 - an installation has finished
 - an uninstallation has finished

The condition is initialized the first time the reconcile loop is run
and a new kataconfig CR instance was found.

The Upgradeable condition can be overriden by an administrator:
https://olm.operatorframework.io/docs/advanced-tasks/communicating-operator-conditions-to-olm/#overriding-operatorconditions

Signed-off-by: Jens Freimann <jfreimann@redhat.com>