switch webhook certificate management from cert-manager to OLM
An operator-sdk-generated webhook is configured by default to rely on
cert-manager for its certificate management.  However, cert-manager is not
a standard part of OpenShift clusters.  To avoid depending on it we take
advantage of the fact that Operator Lifecycle Manager can do certificate
management for webhooks as well.

Note that the switch from cert-manager to OLM happens just by declaring the
webhook to the OLM in ClusterServiceVersion - no need to change the
cert-manager-specific parts of previous commits (mainly in
config/default/kustomization.yaml).

As for the webhook declaration in CSV, it essentially duplicates
information from the ValidatingWebhookConfiguration resource that's used to
declare the webhook to Kubernetes control plane.  The output of
`${KUSTOMIZE} build config/default` was helpful in getting that
information at one place.