KATA-1011: update Status of nodes with selected label
We used to select nodes labelled as workers only. This is wrong
when the user selected a custom label for nodes where the runtime
should be installed. Another case where this doesn't work is on
compact clusters because there we use the master machine config pool.

This can be fixed by using the label that is specified in the
KataConfigPoolSelector.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>