Introduce metrics dashboard in OCP console.
Ading a dashboard in OCP requires creating a ConfigMap under the
openshift-config-managed namespace.
We can't directly install it as a YAML file, as OLM will apply the operator's
namespace systematically.
To work around that, the ConfigMap is created under the the operator's namespace
first, and when kataconfig is deployed, it is copied and installed under the
appropriate namespace.

The resulting dashboard is accessible under the "Observe/Dashboard" section
in the OpenShift console.

Signed-off-by: Julien Ropé <jrope@redhat.com>