enable changing CRI-O log level on kata workers by changing KataConfig
To debug some kinds of problems with kata-enabled workers it's helpful
or even necessary to make CRI-O on the workers log in debug mode.  So
far, this was in principle achievable by a user but some knowledge of
internals of Sandboxed Containers Operator and OpenShift itself was
required.

This commit introduces a new 'logLevel' field in KataConfig.spec which
can be used to set easily a corresponding log level on nodes.
Internally, it works by creating or updating a ContainerRuntimeConfig
instance which requests MCO to set the desired log level on relevant
nodes.  The relevant nodes are all masters (more specifically, all
members of the master Machine Config Pool) in case of a converged
cluster, or nodes defined by the OSC-internal 'kata-oc' MCP otherwise.

To support this implentation, this commit gives the OSC operator a
permission to manipulate ContainerRuntimeConfig resources and labels the
'kata-oc' MCP to enable its selection in
ContainerRuntimeConfig.spec.machineConfigPoolSelector.