KATA-1340: allow re-install on all workers
When we first install with KataConfigPoolSelector set on specific nodes
and later on remove that selector from kataconfig, re-installation on
all worker nodes is never going to happen.

The reason is that the kata-oc machine config pool is never updated
correctly, the old matchLabel key-value pair is never removed.

This patch fixes the problem by adding all matchLabel entries when
kataConfigPoolSelector is non-empty in every reconcile run.  When the
selector is deleted from kataconfig we create a new kata-oc machine
config pool with only our default matchLabels for the worker and kata-oc
role set. The kata-oc entry is required because we always create this
role even when we install on all worker nodes.

The deletetion-and-reinstall also works when a matchExpression is used
in kataconfig instead of matchLabels. This is because machine config
pools nodeSelector always works with matchLabels only.

Fixes: https://issues.redhat.com/browse/KATA-1340
Signed-off-by: Jens Freimann <jfreimann@redhat.com>