kata-monitor DaemonSet: add the "listen-address" arg
In the upstream kata-containers repo the kata-monitor binary is now
listening on the loopback address by default:
https://github.com/kata-containers/kata-containers/commit/e64c54a2ad7

Since the DaemonSet needs to expose the endpoint to Prometheus, let's
specify the "listen-address" argument.
This will cause no changes with older/current kata-monitor images but
will allow to switch seamlessy to the latest kata-monitor images
preserving the availability of the endpoint to Prometheus.

Signed-off-by: Francesco Giudici <fgiudici@redhat.com>