Move SELinuxOptions to kata-monitor
We currently deploy a custom SELinux type on worker nodes and have the
monitor PODs using it through a custom SCC. This is wrong because an
SCC can potentially be applied to any POD anywhere. This is exactly
what happens with the MCO : it exclusively runs on master nodes where
our custom SELinux type is unavailable but the SCC admission logic
legitimately applies our SCC and the MCO fails to start.

Relax the SELinux contstraint in our SCC so that it can work with
any context, and have the monitor requiring explicitely the SELinux
type it wants to use.

Fixes: KATA-1373

Suggested-by: Stanislav Laznicka <slaznick@redhat.com>
Signed-off-by: Greg Kurz <groug@kaod.org>
(cherry picked from commit d4745883e38f287a3d9f934f472ed7aa468f5833)
Signed-off-by: Greg Kurz <groug@kaod.org>