basic setup of peerpod conroller
Peer pods are pods that run in a remote environment instead of
the local cluster that the operator is running on.

Since this is a rather large and distinct new feature we are
adding a second controller. The peer pod controller is imported
to the current kataconfig controller. Support is enabled by adding
a flag in the kataconfig CRD in the next patch.

To use the peer pod controller in this project we need to import it
and give the kataconfig controller access rights to access
(get/list/create/delete/update) instances of the peer pod config
custom resource.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>