set env variable for target runtimeclass
We want to be able to change the target runtime class which is used
by the peer pod webhook that mutates pods with a certain runtime class.

By setting the environment variable in the manager pod it can be
accessed by the webhook which is running in the same process.

Other ways of retrieving the value would be to add it to the CRD.
However this would increase the latency added by the webhook. Given that
it is called for every Pod create/update/delete we need to avoid that.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>