Set params of the controller Deployment's rolling update
This shouldn't be strictly necessary but it turned out that the Deployment
doesn't work with the default settings - instead of rolling out a new
controller version, a new version pod will just start and terminate
immediately.

Without the controller Deployment working properly, full functionality of
storing kata-monitor image URL in an env var cannot be achieved.  More
specifically, cluster users wouldn't be able to specify their own
kata-monitor image as this is done by editing the manager pod template in
the controller-manager Deployment and relies on a subsequent correct
update.