Add security contexts to the operator
OCP 4.12 is now enforcing the "restricted" Pod Security Admission (PSA)
policy by default for every pod. This includes the pods that belong
to the operator, i.e. the controller manager and the monitors. As a
consequence, subscribing to OSC currently fails with the following
error:

      message: 'pods "controller-manager-6dff96dbb8-m2twk" is forbidden: violates
        PodSecurity "restricted:latest": allowPrivilegeEscalation != false (container
        "manager" must set securityContext.allowPrivilegeEscalation=false), unrestricted
        capabilities (container "manager" must set securityContext.capabilities.drop=["ALL"]),
        runAsNonRoot != true (pod or container "manager" must set securityContext.runAsNonRoot=true),
        seccompProfile (pod or container "manager" must set securityContext.seccompProfile.type
        to "RuntimeDefault" or "Localhost")'

The controller manager has two containers, the manager itself and the
kube-rbac-proxy sidecar. None of these two require elevated privileges.
The operator manifest should thus have appropriate security contexts
that explicitly comply with the "restricted" policy as explained here [1].

The monitors do require some privileges though. This is already addressed
with the current code base : the controller adds the appropriate labels
to the OSC namespace. This tells the PSA that it should accept "privileged"
pods to be started in the OSC namespace.

[1] https://sdk.operatorframework.io/docs/best-practices/pod-security-standards/#what-does-that-mean

Fixes: https://issues.redhat.com/browse/KATA-1754
Signed-off-by: Greg Kurz <groug@kaod.org>