generate files in dockerfiles (#219)
These changes add multistage docker files for both the operator and the
operator bundle images. The builder stage uses the make targets to
generate the necessary files to build the two images. This should fix
the Prow CI image build issues and it should allow the copy of the
bundle files in the midstream repo to be removed. Ultimately it means
the same process is being use to build the images no matter what is
building them.

Original patch by Cameron Meadors <cmeadors@redhat.com>
Backported by

Signed-off-by: Jens Freimann <jfreimann@redhat.com>