add runtimeclass.yaml to kustomize file in config/peerpods
I added the runtimeclass.yaml file in an earlier patch but forgot
to add it to kustomization.yaml. This patches fixes it.

Before this commit an error was shown when trying to start a peer pod

Error from server (Forbidden): error when creating "nginx.yaml": pods
"nginx" is forbidden: pod rejected: RuntimeClass "kata-remote-cc" not
found

Signed-off-by: Jens Freimann <jfreimann@redhat.com>