introduce Node watching
As it turns out we cannot reliably handle Node labelling changes without
watching Nodes explicitly.  Notably, labelling a new Node to add it to a
kata-enabled nodes pool will not work (KATA-1928[*]) since this controller
will never learn about it.

The immediate reason is that the controller only watches MCPs but no MCP
is changed.  In particular the "kata-oc" MCP doesn't change since just
adding labels so that a Node matches KataConfig.spec.kataConfigPoolSelector
doesn't yet make it a member of the "kata-oc" MCP because the MCP
membership also requires the "kata-oc" node role label on the Node.  The
node role is supposed is to added (and managed) by this controller yet it
cannot do that since there's nothing to trigger the appropriate
reconciliation.

To fix this, this change introduces Node resource watching.  This enables
us to trigger reconciliation on relevant Node resource changes.  More
specifically we only trigger reconciliation when the changed Node is a
worker.

On creation we reconcile if the new Node matches the kataConfigPoolSelector
on the assumption that non-matching Nodes are of no interest to us.

We ignore Node deletion since it's not clear what good a reconciliation
could do with a Node resource that's already unavailable.

On update we check for relevant changes in the Node resource, namely the
machineconfiguration.openshift.io/state annotation (which is used by
reconciliation code) and labelling changes that change the Node's
membership in kataConfigPoolSelector.

The addition of getNodeSelectorAsMapNoKataOc() should be considered
temporary, pending a future refactoring of kata node selector handling.

[*] https://issues.redhat.com/browse/KATA-1928

Signed-off-by: Pavel Mores <pmores@redhat.com>