introduce new labelling helper functions
This commit just prepares the ground, introducing new functionality but
actually changing nothing yet.

updateNodeLabels() is, in a way, a generalisation of the existing helper
labelNodes().  In addition to what labelNodes() already does, i.e. put
the "kata-oc" node role on every Node that matches
kataConfigPoolSelector, the new updateNodeLabels() function also ensures
that Nodes which do not match the kata selector (anymore) do not have
the "kata-oc" node role either.  It thus adds a critical piece of
functionality that was missing from the controller code and whose lack
made it impossible to manage the "kata-oc" node role label correctly.

unlabelNodes() does what unlabelNode() has been doing but for all
pertinent Nodes.  Due to the way Golang error handling works, it
unfortunately seems impractical for unlabelNodes() to employ
unlabelNode() as its own helper so long as unlabelNode() can return an
error.

Signed-off-by: Pavel Mores <pmores@redhat.com>