fix "kata-oc" node role label handling
This commit changes the "kata-oc" node role label handling as follows.

The "kata-oc" MCP's nodeSelector now only ever contains just a single
requirement - that Node have the "kata-oc" node role label.

Installation procedure now updates node role labels to match the current
kataConfigPoolSelector as the first step.  Label update was pulled out of
newMCPforCR() into processKataConfigInstallRequest() since it's a
first-class concept of the installation algorithm.  It also handles
removing the "kata-oc" role from Nodes that don't match the kata node
selector anymore (see also previous commit).  This is an improvement on the
previous state which only ever added labels during installation.  That was
wrong since kata node selector updates are handled as installations.  This
is true even if they end up narrowing the kata-enabled Node pool and thus
actually uninstalling kata from (a set of) Nodes.

Since the "kata-oc" MCP's nodeSelector is now immutable, code to update it
was removed accordingly from installation handling.

Uninstallation procedure now removes the "kata-oc" node role label from all
Nodes as its first step.  This doesn't really change the semantics of the
algorithm, strictly speaking, but makes for a much more streamlined and
tractable sequence of uninstallation steps.  Unlabeling first drains the
"kata-oc" pool immediately and the rest of uninstallation is just waiting
for individual Nodes to become Updated/Ready in "worker".  The steps used
so far started with removal of the kata extension MC which sent Nodes
updating to a new machine configuration that now matched that of "worker"
but still left them members of the "kata-oc" pool.  Only after individual
Nodes finished updating they were unlabeled one by one and became members
of "worker".  While the end result is the same AFAICT this was a rather
roundabout way and quite a bit harder to follow/make sense of.  This commit
also makes the uninstallation symmetrical to installation from the MCO
standpoint, just with the roles of the source and target pool swapped
between "kata-oc" and "worker".

Signed-off-by: Pavel Mores <pmores@redhat.com>
Fixes: https://issues.redhat.com/browse/KATA-1928