introduce helper functions for two types of kata node selector
KataConfig.spec.kataConfigPoolSelector allows users to install kata on
just a subset of workers by specifying requirements that worker labels
have to match.  The selector's type is metav1.LabelSelector which means
it supports both the MatchLabels and MatchExpressions way of specifying
label requirements.  However, a couple of k8s resources which this
controller creates (namely PodSpec in Daemonset pod template and
RuntimeClass) use node selectors that only support MatchLabels.  The
current controller resolves this mismatch by only passing the
MatchLabels part of kataConfigPoolSelector to the node selectors that
don't support MatchExpressions.  Obviously, ignoring the
MatchExpressions part of kataConfigPoolSelector like this causes
disappointing results for users who opt for MatchExpressions to specify
their kata-enabled worker subset.

To overcome this problem we introduce two kinds of kata node selector.
The first one, called KataConfigNodeSelector in this patch, corresponds
basically to a full set of label requirements coming from a KataConfig
instance (specifically its kataConfigPoolSelector and - if enabled -
checkNodeEligibility members).  This kind of selector is only used
internally in the controller code to label matching nodes with the
"kata-oc" node role.

This node role label then allows us to form the other kind of kata node
selector, called simply NodeSelector here (mostly for historical reasons),
which fits comfortably into MatchLabels and can be passed to any
node-selecting k8s resource without loss of information.

This commit introduces a pair of helper functions for each kind of kata
node selector, one that defines it and one convenience wrapper to convert
it to a commonly used type, respectively.

Signed-off-by: Pavel Mores <pmores@redhat.com>