clean up by removing obsolete functions
getNodeSelectorAsMapNoKataOc() has always been a temporary function to
bridge the gap until the proper solution is in place.

getNodeSelectorAsSelector() is obsolete as NodeSelector is never needed
as labels.Selector.  labels.Selector is a type that, unlike the closely
related LabelSelector, supports actually matching a set of labels against
it.  This is now only done with KataConfigNodeSelector, NodeSelector is
only ever used as a value of k8s resources' nodeSelector fields.

Signed-off-by: Pavel Mores <pmores@redhat.com>