remove KataConfig.status.BaseMcpGeneration and MCP generation comparisons
mcp.status.observedGeneration tracks values of metadata.generation,
specifically it records the last value of metadata.generation that the MCP
controller has processed.  metadata.generation only changes when mcp.spec
changes.  kata-oc MCP's spec is not changed during the OSC controller
normal operation.

The above means that mcp.status.observedGeneration is actually not of
much use for this controller.  With the current OSC the MCP's
observedGeneration does change once since we first create the MCP and
only then we create the extension MC which changes kata-oc's
mcp.spec.configuration.source.  This however is arguably a flaw itself
that is going to be fixed soon as it serves no known purpose and is
confusing because it causes two rendered-kata-oc-* MCs to be created -
one is the legitimate one which includes the extension MC and the other
is spurious and corresponds to the MCP state right after its creation
when the extension MC doesn't yet exist (so this rendered MC is actually
identical to the rendered worker MC).

This commit also removes the MCP retrieval in createExtensionMc() since it
served no other purpose than to get the observedGeneration value.

Signed-off-by: Pavel Mores <pmores@redhat.com>