fix handling of invalid matchExpressions not caught by k8s checking
Before we start actual reconciliation we make sure our node selector is
valid.  This is done indirectly by trying to convert it to labels.Selector
as this conversion conveniently finds any problems with the selector.

We also add logging to make any future bugs in this area easier to diagnose
and fill in selector conversion error handling in updateNodeLabels() which
was originally omitted by mistake.  With the same check done before
reconciliation even starts this is technically superfluous but it seems
better to stay on the safe side.

Signed-off-by: Pavel Mores <pmores@redhat.com>