rework return value of createExtensionMc()
createExtensionMc() now returns the usual tuple of (value,error).

Return value handling has changed so that we requeue with exponential
back-off on error and requeue in 5 seconds in case the MC was just created.
The latter will be useful shortly when we move the MC creation before the
MCP creation to make sure the MC already exists in the cluster when we
create the MCP.  This will prevent creation of spurious rendered-kata-oc-*
MC that's identical to the worker machine configuration.

Error handling and logging was also enhanced within the function.

Signed-off-by: Pavel Mores <pmores@redhat.com>