create extension MC before creating kata-oc MCP
Creating an MC before an MCP that depends on it makes for a more
straightforward sequence of steps and avoids rendering a spurious
rendered-kata-oc-* MC which corresponds to the kata-oc machine
configuration after kata-oc creation and before the extension MC creation.

For this to work though, a (beginning of a) getMcpName() refactor is also
included in this commit.  Curiously enough, the original getMcpName()
implementation won't report "kata-oc" unless that MCP has already been
created in a cluster.  This original implementation has been kept to
preserve compatibility with other existing callers but renamed to
getMcpNameIfMcpExists().

The new name is admittedly clumsy but it describes better the actual
semantics which seem rather clumsy, too.  More importantly, it is expected
that we only keep this function temporarily, until the remaining callers
are refactored to use the new one.  There are two of them and none of them
should be exceedingly hard to adapt(*).

A conditional in processKataConfigInstallRequest() was changed to check
directly for "kata-oc".  Now that getMcpName() can only ever return one of
just two values there's no reason anymore to check indirectly.

For converged clusters this commit should make no difference.

(*) processKataConfigDeleteRequest() doesn't seem to rely on the original
getMcpName()'s kata-oc existence checking. On deletion, kata-oc does exist
pretty much by definition, moreover the function's implementation doesn't
even seem to handle the case where it didn't, apart from just requeueing.
The other caller, getMcp(), goes on to retrieve the MCP from cluster so
it actually already does its own existence check as is.

Signed-off-by: Pavel Mores <pmores@redhat.com>